+++ src/main/java/dev/runabout/RunaboutServiceBuilder.java
@@ -1,7 +1,6 @@
 package dev.runabout;
 
 import java.lang.reflect.Method;
-import java.sql.Timestamp;
 import java.time.Instant;
 import java.util.Objects;
 import java.util.Optional;
@@ -23,7 +22,7 @@ public class RunaboutServiceBuilder<T extends JsonObject> {
     private Set<Class<?>> callerClassBlacklist;
     private RunaboutSerializer customSerializer;
     private Consumer<Throwable> throwableConsumer;
-    private Supplier<Timestamp> datetimeSupplier;
+    private Supplier<String> datetimeSupplier;
     private Function<Method, String> methodToStringFunction;
     private Predicate<StackWalker.StackFrame> stackFramePredicate;
     private RunaboutAPIBuilder emitterBuilder;
@@ -123,13 +122,13 @@ public class RunaboutServiceBuilder<T extends JsonObject> {
 
     /**
      * Sets the datetime supplier for the RunaboutService.
-     * The supplier should return the current datetime UTC as a SQL Timestamp formatted datetime.
-     * By default, the supplier will return the current datetime wrapped in a {@link Timestamp}.
+     * The supplier should return the current datetime UTC as an ISO-8601 formatted string.
+     * By default, the supplier will return the current datetime using {@link Instant#now()}.
      *
      * @param datetimeSupplier The datetime supplier.
      * @return The RunaboutServiceBuilder.
      */
-    public RunaboutServiceBuilder<T> setDatetimeSupplier(Supplier<Timestamp> datetimeSupplier) {
+    public RunaboutServiceBuilder<T> setDatetimeSupplier(Supplier<String> datetimeSupplier) {
         this.datetimeSupplier = datetimeSupplier;
         return this;
     }
@@ -209,8 +208,8 @@ public class RunaboutServiceBuilder<T extends JsonObject> {
         final Consumer<Throwable> throwableConsumerFinal = Optional.ofNullable(throwableConsumer)
                 .orElse(RunaboutServiceBuilder::defaultThrowableConsumer);
 
-        final Supplier<Timestamp> datetimeSupplierFinal = Optional.ofNullable(this.datetimeSupplier)
-                .orElseGet(() -> () -> Timestamp.from(Instant.now()));
+        final Supplier<String> datetimeSupplierFinal = Optional.ofNullable(this.datetimeSupplier)
+                .orElseGet(() -> () -> Instant.now().toString());
 
         final Function<Method, String> methodToStringFunctionFinal = Optional.ofNullable(this.methodToStringFunction)
                 .orElse(RunaboutUtils::methodToRunaboutString);